Desktop/AI YouTube Script Genertaed/backend/requirements.txt
alembic==1.13.3
fastapi==0.115.5
httpx==0.27.2
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
python-jose==3.3.0
SQLAlchemy==2.0.36
uvicorn[standard]==0.32.0

